backend/requirements.txt
fastapi>=0.124.1
uvicorn==0.27.0
google-genai==0.3.0
google-adk==1.27.1
python-dotenv==1.0.0
firebase-admin==6.4.0
google-cloud-storage==2.16.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
